:recycle: re-factoring routes with blueprint
